requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
notionhq-client==0.6.0
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.1
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==73.0.1
simplejson==3.19.3
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2